AVX-136: Change detection of color to brightness of IR
This change replaces the use of hue/color with only pixel brightness
for use with infrared emitters.

Settings are no longer configurable and the button doesn't put the
camera in a different mode. No additional configuration is required
after programming the firmware.

Flash and memory space usage have been freed up for future addition
of logging to SD Card.